// newcode/libs/modules/basics/costcodes/src/lib/model/interfaces/basics-cost-codes-controlling-cost-codes.interface.ts
/*
 * Copyright(c) RIB Software GmbH
 */

export interface IControllingCostCodes {
	/*
	 * showAssignToChildren
	 */
    ShowAssignToChildren?: boolean;

	/*
	 * showOverwrite
	 */

	ShowOverwrite?: boolean;

	/*
	 * showPreserveSelection
	 */
    ShowPreserveSelection: boolean;

	/*
	 * isOverwrite
	 */
    IsOverwrite: boolean;
	/*
	 * isAssignToChildren
	 */

	IsAssignToChildren?: boolean;

	/*
	 * isPreserveSelection
	 */
	IsPreserveSelection?: boolean;
}